pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open_telco"
version = "0.1.0"
description = "Open Telco reasoning benchmarks"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "inspect-ai",
    "openai",
    "datasets",
    "python-dotenv",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = [
    "open_telco",
    "open_telco.telelogs",
    "open_telco.telemath",
    "open_telco.teleqna",
    "open_telco.scripts",
]

